requirements.txt
flask
flask_cors
Flask-SQLAlchemy==3.0.2
psycopg2
pyyaml==5.4.1